+++ nautobot_golden_config/tests/test_api.py
@@ -6,6 +6,8 @@ from rest_framework import status
 
 from nautobot.utilities.testing import APITestCase
 
+from .conftest import create_device, create_feature_rule_json, create_config_compliance
+
 
 User = get_user_model()
 
@@ -13,10 +15,69 @@ User = get_user_model()
 class GoldenConfigAPITest(APITestCase):  # pylint: disable=too-many-ancestors
     """Test the ConfigCompliance API."""
 
+    def setUp(self):
+        """Create a superuser and token for API calls."""
+        super().setUp()
+        self.device = create_device()
+        self.compliance_rule_json = create_feature_rule_json(self.device)
+        self.base_view = reverse("plugins-api:nautobot_golden_config-api:configcompliance-list")
+
+    def test_root(self):
+        """Validate the root for Nautobot Chatops API."""
+        url = reverse("plugins-api:nautobot_golden_config-api:api-root")
+        response = self.client.get("{}?format=api".format(url), **self.header)
+        self.assertEqual(response.status_code, 200)
+
     def test_device_list(self):
         """Verify that devices can be listed."""
         url = reverse("dcim-api:device-list")
         self.add_permissions("dcim.view_device")
         response = self.client.get(url, **self.header)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["count"], 0)
+        self.assertEqual(response.data["count"], 1)
+
+    def test_config_compliance_list_view(self):
+        """Verify that config compliance objects can be listed."""
+        actual = '{"foo": {"bar-1": "baz"}}'
+        intended = '{"foo": {"bar-2": "baz"}}'
+        create_config_compliance(
+            self.device, actual=actual, intended=intended, compliance_rule=self.compliance_rule_json
+        )
+        self.add_permissions("nautobot_golden_config.view_configcompliance")
+        response = self.client.get(self.base_view, **self.header)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["count"], 1)
+
+    def test_config_compliance_post_new_json_compliant(self):
+        """Verify that config compliance detail view."""
+        self.add_permissions("nautobot_golden_config.add_configcompliance")
+        response = self.client.post(
+            self.base_view,
+            data={
+                "device": self.device.id,
+                "intended": '{"foo": {"bar-1": "baz"}}',
+                "actual": '{"foo": {"bar-1": "baz"}}',
+                "rule": self.compliance_rule_json.id,
+            },
+            format="json",
+            **self.header,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertTrue(response.data["compliance"])
+
+    def test_config_compliance_post_new_json_not_compliant(self):
+        """Verify that config compliance detail view."""
+        self.add_permissions("nautobot_golden_config.add_configcompliance")
+        response = self.client.post(
+            self.base_view,
+            data={
+                "device": self.device.id,
+                "intended": '{"foo": {"bar-1": "baz"}}',
+                "actual": '{"foo": {"bar-2": "baz"}}',
+                "rule": self.compliance_rule_json.id,
+            },
+            format="json",
+            **self.header,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertFalse(response.data["compliance"])