+++ nautobot_golden_config/tests/test_utilities/test_git.py
@@ -14,11 +14,13 @@ class GitRepoTest(unittest.TestCase):
         mock_obj.filesystem_path = "/fake/path"
         mock_obj.remote_url = "/fake/remote"
         mock_obj._token = "fake token"  # pylint: disable=protected-access
+        mock_obj.username = None
         self.mock_obj = mock_obj
 
     @patch("nautobot_golden_config.utilities.git.Repo", autospec=True)
     def test_gitrepo_path_noexist(self, mock_repo):
         """Test Repo is not called when path isn't valid, ensure clone is called."""
+        self.mock_obj.username = None
         GitRepo(self.mock_obj)
         mock_repo.assert_not_called()
         mock_repo.clone_from.assert_called_with("/fake/remote", to_path="/fake/path")
@@ -28,6 +30,35 @@ class GitRepoTest(unittest.TestCase):
     def test_gitrepo_path_exist(self, mock_repo, mock_os):
         """Test Repo is not called when path is valid, ensure Repo is called."""
         mock_os.path.isdir.return_value = True
+        self.mock_obj.username = None
         GitRepo(self.mock_obj)
         mock_repo.assert_called_once()
         mock_repo.assert_called_with(path="/fake/path")
+
+    @patch("nautobot_golden_config.utilities.git.os")
+    @patch("nautobot_golden_config.utilities.git.Repo", autospec=True)
+    def test_path_exist_token_and_username(self, mock_repo, mock_os):
+        """Test Repo is not called when path is valid, ensure Repo is called."""
+        mock_os.path.isdir.return_value = True
+        self.mock_obj.username = "Test User"
+        GitRepo(self.mock_obj)
+        mock_repo.assert_called_once()
+        mock_repo.assert_called_with(path="/fake/path")
+
+    @patch("nautobot_golden_config.utilities.git.os")
+    @patch("nautobot_golden_config.utilities.git.Repo", autospec=True)
+    def test_username_with_symbols(self, mock_repo, mock_os):
+        """Test Repo is not called when path is valid, ensure Repo is called."""
+        mock_os.path.isdir.return_value = True
+        self.mock_obj.username = "user@fakeemail.local"
+        GitRepo(self.mock_obj)
+        mock_repo.assert_called_once()
+        mock_repo.assert_called_with(path="/fake/path")
+
+    @patch("nautobot_golden_config.utilities.git.os")
+    @patch("nautobot_golden_config.utilities.git.Repo", autospec=True)
+    def test_git_with_username(self, mock_repo, mock_os):  # pylint: disable=unused-argument
+        """Test username with special character works."""
+        self.mock_obj.username = "admin@ntc.com"
+        GitRepo(self.mock_obj)
+        mock_repo.assert_called_once()